marketing/SEO/requirements.txt
requests==2.32.3
pyyaml==6.0.2
anthropic==0.40.0
responses==0.25.3
pytest==8.3.4
python-dotenv==1.0.1
